+++ lib/http/cache/headers.rb
@@ -69,9 +69,9 @@ module HTTP
       # ---
       # Some servers send a "Expire: -1" header which must be treated as expired
       def seconds_til_expires
-        get("Expires")
-          .map { |e| http_date_to_ttl(e) }
-          .max
+        get("Expires").
+          map { |e| http_date_to_ttl(e) }.
+          max
       end
 
       def http_date_to_ttl(t_str)
@@ -89,11 +89,11 @@ module HTTP
 
       # @return [Numeric] the value of the max-age component of cache control
       def explicit_max_age
-        get("Cache-Control")
-          .map { |v| (/max-age=(\d+)/i).match(v) }
-          .compact
-          .map { |m| m[1].to_i }
-          .max
+        get("Cache-Control").
+          map { |v| (/max-age=(\d+)/i).match(v) }.
+          compact.
+          map { |m| m[1].to_i }.
+          max
       end
     end
   end